# Parser crude optimizations.