@quetz 2018-03-20
I rebased it off master so that #144 fix gets in