@gunnsth 2018-03-20
Ran this on a basic pass-through (read-write) test corpus with 15006 files.  About 1.1% reduction in total run time.  Same number of errors on the corpus.  Looks good.